feat: add layout component, change husky config
