+++ .husky/pre-push
@@ -1,1 +1,1 @@
-npm run test
+#!/npm run test
