+++ src/components/themeBtn/ThemeBtn.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from '../../contexts/themeContext';
+import { ThemeContext } from '../../contexts/themeProvider';
 import classes from './ThemeBtn.module.scss';
 
 const ThemeBtn = () => {
